# app/openapi.yml
openapi: "3.0.0"
servers:
  - description:  Test
    url: /
info:
  description: This is a simple API
  version: "1.0.0"
  title: Test API
paths:
...
  /dealer_inventory:
    post:
      tags:
        - Dealer_Inventory
      summary: Add a new Car inventory record for dealer 
      operationId: app.views.dealerInventoryView.add_dealer_inventory
      description: |
        Dealer Inventory is the associative mapping between a dealer and Car. It lists the number of cars present in inventory for that dealer. 
        Adds a new Car Inventory record for dealer 
      requestBody:
        description: Creates a new car inventory record for dealer  
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealerInventory'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags:
        - Dealer_Inventory
      summary: List all inventory records for all dealers  
      operationId: app.views.dealerInventoryView.get_dealer_inventory
      description: |
        List all inventory records for all dealers  
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllDealerInventoryResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /dealer_inventory/{dealer_id}:
    get:
      tags:
        - Dealer_Inventory
      summary: List all cars and their respective counts for a particular Dealer ID 
      operationId: app.views.dealerInventoryView.get_dealer_inventory_by_dealer_id
      description: |
        List all cars and their respective counts for a particular Dealer ID 
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllDealerInventoryResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /dealer_inventory/{dealer_id}/{car_serial_no}:
    get:
      tags:
        - Dealer_Inventory
      summary: List Car Count in inventory for a particular dealer and car serial number combination
      operationId: app.views.dealerInventoryView.get_dealer_inventory_by_dealer_car
      description: |
        List Car Count in inventory for a particular dealer and car serial number combination
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
        - name: car_serial_no
          in: path 
          description: Car Serial Number
          required: true 
          schema:
            type: string 
            example: 'C-1'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDealerInventoryResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'


    patch:
      tags:
        - Dealer_Inventory
      summary: Update/Patch Dealer Inventory by dealer id and car serial number 
      operationId: app.views.dealerInventoryView.update_dealer_inventory
      description: |
        Updates a single record based on dealer_id and car_serial_no 
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
        - name: car_serial_no
          in: path 
          description: Car Serial Number
          required: true 
          schema:
            type: string 
            example: 'C-1'
      requestBody:
        description: Updates a dealer inventory record
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealerInventoryPatch'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Dealer_Inventory
      summary: Deletes Dealer Inventory by dealer id and car serial number 
      operationId: app.views.dealerInventoryView.delete_inventory
      description: |
        Deletes a single dealer inventory record 
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
        - name: car_serial_no
          in: path 
          description: Car Serial Number
          required: true 
          schema:
            type: string 
            example: 'C-1'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /dealer:
    post:
      tags:
        - Dealer
      summary: Add a new Dealer
      operationId: app.views.dealerView.add_dealer
      description: |
        Add a new Dealer
      requestBody:
        description: Creates a new dealer 
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dealer'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags:
        - Dealer
      summary: List all dealers
      operationId: app.views.dealerView.get_dealer
      description: |
        List all dealers
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllDealerResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /dealer/{dealer_id}:
    get:
      tags:
        - Dealer
      summary: Get dealer by dealer_id 
      operationId: app.views.dealerView.get_dealer_by_id
      description: |
        Get dealer by dealer_id 
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDealerResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  
    patch:
      tags:
        - Dealer
      summary: Update/Patch Dealer by dealer id
      operationId: app.views.dealerView.update_dealer
      description: |
        Updates a single record based on dealer_id 
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
      requestBody:
        description: Updates a dealer inventory record
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealerPatch'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Dealer
      summary: Deletes Dealer by dealer id
      operationId: app.views.dealerView.delete_dealer
      description: |
        Deletes a single dealer  record 
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /dealer/state/{state}:
    get:
      tags:
        - Dealer
      summary: Get all dealers by state
      operationId: app.views.dealerView.get_dealer_state
      description: |
        Get all dealers by state
      parameters:
        - name: state
          in: path 
          description: State to filter by 
          required: true 
          schema:
            type: string 
            example: california
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllDealerResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /state_tax/state/{state}:
    get:
      tags:
        - State_Tax
      summary: List all State_Tax items for a given state 
      operationId: app.views.stateTaxView.get_state_tax
      description: |
        List all State_Tax items for a given state 
      parameters:
        - name: state
          in: path 
          description: state
          required: true 
          schema:
            type: string 
            example: california
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStateTaxResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  schemas:
...
    DealerInventory:
      required: 
        - car_serial_no
        - dealer_id
        - car_count 
      type: object
      properties:
        car_serial_no:
          type: string
          example: C-1
        dealer_id:
          type: string 
          example: 1
        car_count:
          type: integer 
          example: 100
    GetAllDealerInventoryResponse:
      type: object 
      properties:
        msg:
          type: array
          items:
            $ref: '#/components/schemas/DealerInventoryObject'
        status:
          type: string 
    DealerInventoryObject:
      type: object
      properties:
        dealer_inventory_id:
          type: string
          example: 1
        car_serial_no:
          type: string
          example: C-1
        dealer_id:
          type: string 
          example: 1
        car_count:
          type: integer 
          example: 100
        created_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT  
        modified_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
    GetDealerInventoryResponse:
      type: object 
      properties:
        msg:
          $ref: '#/components/schemas/DealerInventoryObject'
        status:
          type: string 
          example: success
    DealerInventoryPatch:
      type: object
      properties:
        car_count:
          type: integer 
          example: 100
    Dealer:
      required: 
        - name
        - web_url
        - street_address 
        - postal_code
        - city
        - state
      type: object
      properties:
        name:
          type: string
          example: Cartec Honda
        web_url:
          type: string 
          example: "cartechonda@example.com"
        street_address:
          type: string 
          example: Abc Street, CA-94301
        postal_code:
          type: string 
          example: 94301
        city:
          type: string 
          example: Palo Alto
        state:
          type: string 
          example: california
    GetAllDealerResponse:
      type: object 
      properties:
        msg:
          type: array
          items:
            $ref: '#/components/schemas/DealerObject'
        status:
          type: string 
    DealerObject:
      type: object
      properties:
        dealer_id:
          type: integer
          example: 1 
        name:
          type: string
          example: Cartec Honda
        web_url:
          type: string 
          example: "cartechonda@example.com"
        street_address:
          type: string 
          example: Abc Street, CA-94301
        postal_code:
          type: string 
          example: 94301
        city:
          type: string 
          example: Palo Alto
        state:
          type: string 
          example: california
        created_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT  
        modified_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
    GetDealerResponse:
      type: object 
      properties:
        msg:
          $ref: '#/components/schemas/DealerObject'
        status:
          type: string 
          example: success
    DealerPatch:
      type: object
      properties:
        name:
          type: string
          example: Cartec Honda
        web_url:
          type: string 
          example: "cartechonda@example.com"
        street_address:
          type: string 
          example: Abc Street, CA-94301
        postal_code:
          type: string 
          example: 94301
        city:
          type: string 
          example: Palo Alto
        state:
          type: string 
          example: california
...
    StateTaxObject:
      type: object
      properties:
        state_tax_id:
          type: integer
          example: 1 
        state:
          type: string 
          example: california
        tax: 
          type: number 
          example: 10 
        created_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT  
        modified_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
    GetStateTaxResponse:
      type: object 
      properties:
        msg:
          $ref: '#/components/schemas/StateTaxObject'
        status:
          type: string 
          example: success
...
    ApiResponse:
        type: object
        properties:
          msg:
            type: string
          status:
            type: string 